refactor(background): align engine slice with FSD
- Move GPUContext to shared/lib/webgpu
- Move simulation core, passes, shaders into features/background
- Delete features/engine slice
- Update AGENTS.md and vite.config.ts exclusions
