+++ src/features/background/model/MarchingSquares.ts
