+++ src/features/background/model/Simulation.ts
