+++ src/features/background/model/SimulationConstants.ts
