+++ src/features/background/model/simulation.worker.ts
