+++ vite.config.ts
@@ -109,8 +109,8 @@ export default defineConfig(({ mode }) => ({
         'src/entities/codeTheme/codeTheme.ts',
         'src/entities/post/api.server.ts',
         'src/entities/post/post.ts',
-        'src/features/engine/passes/IPass.ts',
-        'src/features/engine/core/SimulationState.ts',
+        'src/features/background/lib/passes/IPass.ts',
+        'src/features/background/model/SimulationState.ts',
       ],
     },
   },
